신청자관리 VO, SQL
