+++ jjack/src/main/java/com/jjack/web/applicantList/dao/ApplicantListDAO.java
@@ -1,0 +1,6 @@
+package com.jjack.web.applicantList.dao;
+import org.mybatis.spring.support.SqlSessionDaoSupport;
+
+public class ApplicantListDAO extends SqlSessionDaoSupport{
+
+}
