+++ jjack/src/main/resources/sql/ApplicantListDAO.xml
@@ -1,0 +1,92 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC 	"-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<!-- namespace는 연결할 DAO 인터페이스경로.클래스명 -->	
+<mapper namespace="com.jjack.web.applicantList.dao.ApplicantListDAO">
+	<sql id = "apList">
+		SELECT
+			a.m_no AS mno,
+			a.a_no AS ano,
+			m.m_no AS id,
+			m.name AS name,
+			g.g_nick AS nick,
+			m.m_sexAS sex,
+			m.m_age AS age,
+			m.m_loc AS loc,
+			a.a_interest AS interest,
+			a.apdate_AS adate,
+			a.a_tel AS tel,
+			m.m_email AS email,
+			a.a_cond AS thisStep,
+			a.a_pic AS pic,
+			ac.acount AS acount,
+			gc.gcount AS gcount
+		FROM
+			TB_Member m,
+			TB_Apply a,
+			TB_Guest g,
+			(SELECT COUNT(a_no) AS acount, a_no FROM TB_APPLY) ac,
+			(SELECT COUNT(g_no) AS gcount, g_no FROM TB_GUEST) gc
+		WHERE
+			ac.a_no = a.a_no 
+			AND gc.g_no = g.g_no 
+			AND a.e_eventdate = #{edate} 
+			AND
+	</sql>
+
+	<!-- 승인대기중인 입소신청자들 -->
+	<select id="getWaitList" parameterType="int" resultType="applicantlistvo">
+		<include refid="apList"></include>
+			a_cond = 1
+		ORDER BY
+			a_apdate ASC
+	</select>
+	
+	<!-- 승인된 남자신청자들 -->
+	<select id="getOkManList" parameterType="int" resultType="applicantlistvo">
+		<include refid="apList"></include>
+			a_cond IN (2, 3, 4, 8, 9, 10)
+		AND m_sex = 'M'
+			ORDER BY
+		a_apdate ASC
+	</select>
+	
+	<!-- 승인된 여자신청자들 -->
+	<select id="getOkWomanList" parameterType="int" resultType="applicantlistvo">
+		<include refid="apList"></include>
+			a_cond IN (2, 3, 4, 8, 9, 10)
+		AND m_sex = 'W'
+			ORDER BY
+		a_apdate ASC
+	</select>
+	
+	<!-- 승인거부된 신청자들 -->
+	<select id="getDeniedList" parameterType="int" resultType="applicantlistvo">
+		<include refid="apList"></include>
+			a_cond IN (5,6,7)
+		ORDER BY
+			a_apdate ASC
+	</select>
+	
+	<!-- 선택한 입소날짜, 기수와 대기자, 승인자, 거부자 수 -->
+	<select id="getApCountVO" parameterType="String" resultType="applicantcountvo">
+		SELECT
+			e_eventdate AS edate,
+			e_gisoo AS gisoo,
+			e_title AS etitle,
+			waitCount,
+			manCount,
+			womanCount,
+			deniedCount
+		FROM
+			TB_Event,
+			(SELECT COUNT(*) AS waitCount FROM TB_Apply WHERE e_eventdate = #{edate} AND a_cond=1) wc,
+			(SELECT COUNT(*) AS manCount FROM TB_Apply WHERE e_eventdate = #{edate} AND a_cond IN (2, 3, 4, 8, 9, 10)) mc,
+			(SELECT COUNT(*) AS womanCount FROM TB_Apply WHERE e_eventdate = #{edate} AND a_cond IN (2, 3, 4, 8, 9, 10)) fc,
+			(SELECT COUNT(*) AS deniedCount	FROM TB_Apply WHERE e_eventdate = #{edate} AND a_cond IN (5, 6 ,7))  dc
+		WHERE
+			e_eventdate = #{edate}
+	</select>
+	
+	
+	
+</mapper>
